Move Form to a model
