+++ admin/app/components/certification-centers/information-edit.js
@@ -1,89 +1,16 @@
 import Component from '@glimmer/component';
-import { tracked } from '@glimmer/tracking';
-import { buildValidations, validator } from 'ember-cp-validations';
-import { getOwner } from '@ember/application';
-import Object, { action } from '@ember/object';
+import { action } from '@ember/object';
+import { inject as service } from '@ember/service';
 import { types } from '../../models/certification-center';
 
-const Validations = buildValidations({
-  name: {
-    validators: [
-      validator('presence', {
-        presence: true,
-        ignoreBlank: true,
-        message: 'Le nom ne peut pas être vide',
-      }),
-      validator('length', {
-        min: 1,
-        max: 255,
-        message: 'La longueur du nom ne doit pas excéder 255 caractères',
-      }),
-    ],
-  },
-  type: {
-    validators: [
-      validator('presence', {
-        presence: true,
-        ignoreBlank: true,
-        message: 'Le type ne peut pas être vide',
-      }),
-    ],
-  },
-  externalId: {
-    validators: [
-      validator('length', {
-        min: 0,
-        max: 255,
-        message: "La longueur de l'identifiant externe ne doit pas excéder 255 caractères",
-      }),
-    ],
-  },
-  dataProtectionOfficerFirstName: {
-    validators: [
-      validator('length', {
-        min: 0,
-        max: 255,
-        message: 'La longueur du prénom du DPO ne doit pas excéder 255 caractères',
-      }),
-    ],
-  },
-  dataProtectionOfficerLastName: {
-    validators: [
-      validator('length', {
-        min: 0,
-        max: 255,
-        message: 'La longueur du nom du DPO ne doit pas excéder 255 caractères',
-      }),
-    ],
-  },
-  dataProtectionOfficerEmail: {
-    validators: [
-      validator('length', {
-        max: 255,
-        message: "La longueur de l'adresse e-mail du DPO ne doit pas excéder 255 caractères.",
-      }),
-      validator('format', {
-        allowBlank: true,
-        type: 'email',
-        message: "L'adresse e-mail du DPO n'a pas le bon format.",
-      }),
-    ],
-  },
-});
-
-class Form extends Object.extend(Validations) {
-  @tracked name;
-  @tracked externalId;
-  @tracked type;
-  @tracked habilitations;
-}
-
 export default class InformationEdit extends Component {
+  @service store;
   certificationCenterTypes = types;
 
   constructor() {
     super(...arguments);
-    this.form = Form.create(getOwner(this).ownerInjection());
+    this.form = this.store.createRecord('certification-center-form');
+
     this._initForm();
   }
 
@@ -93,7 +20,7 @@ export default class InformationEdit extends Component {
 
   @action
   selectCertificationCenterType(value) {
-    this.form.type = value;
+    this.form.set('type', value ? value.trim() : value);
   }
 
   @action
@@ -115,9 +42,9 @@ export default class InformationEdit extends Component {
       return;
     }
 
-    this.args.certificationCenter.set('name', this.form.name.trim());
-    this.args.certificationCenter.set('externalId', !this.form.externalId ? null : this.form.externalId.trim());
-    this.args.certificationCenter.set('type', this.form.type.trim());
+    this.args.certificationCenter.set('name', this.form.name);
+    this.args.certificationCenter.set('externalId', !this.form.externalId ? null : this.form.externalId);
+    this.args.certificationCenter.set('type', this.form.type);
     this.args.certificationCenter.set('habilitations', this.form.habilitations);
     this.args.certificationCenter.set('dataProtectionOfficerFirstName', this.form.dataProtectionOfficerFirstName);
     this.args.certificationCenter.set('dataProtectionOfficerLastName', this.form.dataProtectionOfficerLastName);
@@ -128,15 +55,7 @@ export default class InformationEdit extends Component {
   }
 
   _initForm() {
-    const {
-      habilitations,
-      name,
-      externalId,
-      type,
-      dataProtectionOfficerFirstName,
-      dataProtectionOfficerLastName,
-      dataProtectionOfficerEmail,
-    } = this.args.certificationCenter.getProperties(
+    const properties = this.args.certificationCenter.getProperties(
       'habilitations',
       'name',
       'externalId',
@@ -145,13 +64,6 @@ export default class InformationEdit extends Component {
       'dataProtectionOfficerLastName',
       'dataProtectionOfficerEmail'
     );
-
-    this.form.name = name;
-    this.form.habilitations = habilitations ? habilitations.toArray() : [];
-    this.form.externalId = externalId;
-    this.form.type = type;
-    this.form.dataProtectionOfficerFirstName = dataProtectionOfficerFirstName;
-    this.form.dataProtectionOfficerLastName = dataProtectionOfficerLastName;
-    this.form.dataProtectionOfficerEmail = dataProtectionOfficerEmail;
+    this.form.setProperties(properties);
   }
 }
